Deprecate mod_spidermonkey
We consider mod_spidermonkey to be completely replaced by mod_v8 (and
more generally by mod_lua) at this point.  We've barely supported
mod_spidermonkey for years.  Everyone should stop using it.

We're keeping this around for awhile in case people are relying on C
extensions to mod_spidermonkey and need time to migrate.  This module
may or may not be part of a v1.4 release depending on whether we find
anyone cares.  It definitely will not be part of the next release
after v1.4.